backend/pyproject.toml
[project]
name = "gymopus"
version = "0.1.0"
description = "Knowledge-constrained fitness workflow agent"
requires-python = ">=3.12"
dependencies = [
    # AI core
    "langgraph>=0.2",
    "langgraph-checkpoint-postgres>=2.0",
    "pydantic-ai>=0.2",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.3",
    "langchain-postgres>=0.0.12",
    # Observability
    "langsmith>=0.2",
    # Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "psycopg[binary]>=3.1",
    "alembic>=1.14",
    # Tools
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
